# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Minecraft Bingo API",
    "description": "API for managing Minecraft Bingo games with both team-based and solo (free-for-all) modes",
    "version": "1.0.0",
    "contact": {
      "email": "fconteo17@gmail.com"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "https://minecraft-bingo-webapp.vercel.app/"
    }
  ],
...
  "paths": {
...
    "/games/start": {
      "post": {
        "tags": ["Games"],
        "summary": "Start a new game",
        "description": "Creates a new game with the specified teams/players and quests",
        "operationId": "startGame",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GameStartRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Game created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Game"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/games/{gameId}": {
      "get": {
        "tags": ["Games"],
        "summary": "Get game by ID",
        "description": "Retrieves a specific game by its ID",
        "operationId": "getGameById",
        "parameters": [
          {
            "name": "gameId",
            "in": "path",
            "description": "ID of the game to retrieve",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Game"
                }
              }
            }
          },
          "404": {
            "description": "Game not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/games/{gameId}/updateQuest": {
      "post": {
        "tags": ["Games"],
        "summary": "Update quest name",
        "description": "Updates the name of a quest in a game",
        "operationId": "updateQuest",
        "parameters": [
          {
            "name": "gameId",
            "in": "path",
            "description": "ID of the game",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["gameId", "questName", "newQuestName"],
                "properties": {
                  "gameId": {
                    "type": "string",
                    "format": "uuid",
                    "description": "ID of the game"
                  },
                  "questName": {
                    "type": "string",
                    "description": "Current name of the quest"
                  },
                  "newQuestName": {
                    "type": "string",
                    "description": "New name for the quest"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Quest updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Game"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Game or quest not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Quest": {
        "type": "object",
        "properties": {
          "name": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": {
                    "type": "string"
                  }
                }
              }
            ],
            "description": "Name of the quest"
          },
          "completedBy": {
            "oneOf": [
              {
                "type": "string",
                "enum": ["Red", "Blue"]
              },
              {
                "type": "string",
                "description": "Player name in solo games"
              }
            ],
            "description": "Team or player that completed the quest"
          },
          "completedByPlayer": {
            "type": "string",
            "description": "Name of the player who completed the quest"
          },
          "completedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the quest was completed"
          }
        },
        "description": "A quest in the bingo game"
      },
      "Game": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier for the game"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "When the game was created"
          },
          "gameType": {
            "type": "string",
            "enum": ["Teams", "Solo"],
            "description": "Type of game (team-based or free-for-all)"
          },
          "teams": {
            "type": "array",
            "description": "Teams participating in the game (for team-based games)",
            "items": {
              "type": "string",
              "enum": ["Red", "Blue"]
            },
            "minItems": 2,
            "maxItems": 2
          },
          "players": {
            "type": "array",
            "description": "Players participating in the game (for solo games)",
            "items": {
              "type": "string"
            }
          },
          "quests": {
            "type": "array",
            "description": "List of quests in the game",
            "items": {
              "$ref": "#/components/schemas/Quest"
            }
          },
          "winner": {
            "type": "string",
            "description": "Team or player that won the game"
          },
          "completedQuests": {
            "type": "object",
            "description": "Map of completed quests by team/player",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          "ranked": {
            "type": "boolean",
            "description": "Whether this was a ranked game"
          }
        }
      },
      "GameStartRequest": {
        "type": "object",
        "required": ["gameType", "quests"],
        "properties": {
          "gameType": {
            "type": "string",
            "enum": ["Teams", "Solo"],
            "description": "Type of game: Teams (Red vs Blue) or Solo (free-for-all)"
          },
          "team1Name": {
            "type": "string",
            "enum": ["Red", "Blue"],
            "description": "Name of the first team (required for Teams game type)"
          },
          "team2Name": {
            "type": "string",
            "enum": ["Red", "Blue"],
            "description": "Name of the second team (required for Teams game type)"
          },
          "players": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "minItems": 2,
            "description": "List of players (required for Solo game type)"
          },
          "quests": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "minItems": 25,
            "maxItems": 25,
            "description": "List of quest names"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          }
        }
      },
...
    }
  }
}